+++ sdk/cosmosdb/cosmos/test/public/integration/aggregates/groupBy.spec.ts
@@ -8,6 +8,7 @@ import { Context } from "mocha";
 
 const options = {
   maxItemCount: 100,
+  disableNonStreamingOrderByQuery: true,
 };
 
 const items = [